requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-hosts==4.0
django-restframework==0.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
phonenumbers==8.12.18
Pillow==8.1.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
